refactor(plugin): adapt for the plugin generation tool
