+++ internal/pkg/plugin/argocd/validate.go
